+++ packages/layout/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 5.2.0
+
+### Patch Changes
+
+- @lowdefy/block-utils@5.2.0
+- @lowdefy/helpers@5.2.0
+
 ## 5.1.0
 
 ### Patch Changes