+++ packages/plugins/operators/operators-dayjs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @lowdefy/operators-dayjs
 
+## 5.2.0
+
+### Patch Changes
+
+- Updated dependencies [73fa2b9]
+- Updated dependencies [1e964c4]
+  - @lowdefy/operators@5.2.0
+  - @lowdefy/helpers@5.2.0
+
 ## 5.1.0
 
 ### Patch Changes